feat: Added fork of rust-msi to repo.
Added a fork of rust-msi to the repo as I'm trying to troubleshoot
miscellaneous MSI errors and more changes might be required.

The one change to the fork I have already made is to add in support for
the PAGE_COUNT field as that was causing the installer to error
instantly.